+++ tasks/options/connect.js
@@ -38,6 +38,7 @@
                             '/api/account/advertisers': 'mocks/advertiser/main.js',
                             '/api/account/customer': 'mocks/customer/main.js',
                             '/api/account/customers': 'mocks/customer/main.js',
+                            '/api/account/policies': 'mocks/policy/main.js',
                             '/api/content/experience': 'mocks/content/main.js',
                             '/api/content/experiences': 'mocks/content/main.js',
                             '/api/content/category': 'mocks/category/main.js',