+++ CHANGELOG.md
@@ -1,10 +1,11 @@
 # ProShop
 
 * *[v0.7.5]*
-  * Advertisers, Customers and Categories lists no longer load every
+  * [DEV]: Advertisers, Customers and Categories lists no longer load every
     entity when displaying results. Results are fetched as users paginate.
     Filtering by column heading and searching by name now loads new
     results from the server.
+  * [FEATURE]: Policy Manager
 
 ## v0.7.4 (June 15, 2015)